+++ admin/tabs/activity/clienterrors/ClientErrorModel.js
@@ -24,7 +24,7 @@ export class ClientErrorModel {
 
     @managed
     gridModel = new GridModel({
-        stateModel: 'xhClientErrorGrid',
+        persistWith: 'xhClientErrorGrid',
         enableColChooser: true,
         enableExport: true,
         exportOptions: {filename: () => `Client Errors ${fmtDate(this.startDate)} to ${fmtDate(this.endDate)}`},