+++ admin/tabs/general/configs/ConfigModel.js
@@ -24,7 +24,7 @@ export class ConfigModel {
 
     @managed
     gridModel = new RestGridModel({
-        stateModel: 'xhConfigGrid',
+        persistWith: 'xhConfigGrid',
         enableColChooser: true,
         enableExport: true,
         store: new RestStore({