+++ admin/tabs/general/users/UserModel.js
@@ -19,7 +19,7 @@ export class UserModel {
 
     @managed
     gridModel = new GridModel({
-        stateModel: 'xhUserGrid',
+        persistWith: 'xhUserGrid',
         enableColChooser: true,
         enableExport: true,
         store: {idSpec: 'username'},