+++ admin/tabs/preferences/PreferencePanel.js
@@ -15,7 +15,7 @@ export const preferencePanel = hoistCmp.factory(
 );
 
 const modelSpec = {
-    stateModel: 'xhPreferenceGrid',
+    persistWith: 'xhPreferenceGrid',
     enableColChooser: true,
     enableExport: true,
     store: {