+++ cmp/grid/GridModel.js
@@ -42,7 +42,7 @@ import {
     pull,
     sortBy
 } from 'lodash';
-import {GridStateModel} from './GridStateModel';
+import {PersistenceModel} from './impl/PersistenceModel';
 import {GridSorter} from './impl/GridSorter';
 
 /**
@@ -71,8 +71,6 @@ export class GridModel {
     selModel;
     /** @member {boolean} */
     treeMode;
-    /** @member {GridStateModel} */
-    @managed stateModel;
     /** @member {ColChooserModel} */
     @managed colChooserModel;
     /** @member {function} */
@@ -136,6 +134,9 @@ export class GridModel {
     /** @private - initial state provided to ctor - powers restoreDefaults(). */
     _defaultState;
 
+    @managed
+    persistenceModel;
+
     /**
      * Is autosizing enabled on this grid?
      *
@@ -157,10 +158,8 @@ export class GridModel {
      *      `store.SummaryRecord` to be populated. Valid values are true/'top', 'bottom', or false.
      * @param {(StoreSelectionModel|Object|String)} [c.selModel] - StoreSelectionModel, or a
      *      config or string `mode` with which to create one.
-     * @param {(Object|string)} [c.stateModel] - Specification for a GridStateModel.  If an object,
-     *      will be assumed to be a config for a GridStateModel. If a string is provided, a default
-     *      model will be created using a local state provider.  If no value is provided, a default
-     *      model will be created using a transient state provider.
+     * @param {PersistenceProvider} [c.persistWith] - PersistenceProvider or a config to create one.
+     * @param {GridPersistenceOptions} [c.persistOptions] - options governing persistence.
      * @param {?string} [c.emptyText] - text/HTML to display if grid has no records.
      *      Defaults to null, in which case no empty text will be shown.
      * @param {(string|string[]|Object|Object[])} [c.sortBy] - colId(s) or sorter config(s) with
@@ -221,10 +220,12 @@ export class GridModel {
         treeMode = false,
         showSummary = false,
         selModel,
-        stateModel = null,
         emptyText = null,
         sortBy = [],
         groupBy = null,
+
+        persistWith,
+        persistOptions,
 
         sizingMode = 'standard',
         showHover = false,
@@ -251,6 +252,7 @@ export class GridModel {
         experimental,
         ...rest
     }) {
+
         this._defaultState = {columns, sortBy, groupBy};
 
         this.treeMode = treeMode;
@@ -272,7 +274,9 @@ export class GridModel {
         });
 
         apiRemoved(rest.contextMenuFn, 'contextMenuFn', 'Use contextMenu instead');
+        apiRemoved(rest.stateModel, 'stateModel', "Use 'persistWith' and 'persistOptions' arguments instead.");
         apiRemoved(exportOptions.includeHiddenCols, 'includeHiddenCols', "Replace with {columns: 'ALL'}.");
+
         throwIf(
             autosizeOptions.fillMode && !['all', 'left', 'right', 'none'].includes(autosizeOptions.fillMode),
             `Unsupported value for fillMode.`
@@ -303,7 +307,7 @@ export class GridModel {
 
         this.colChooserModel = enableColChooser ? this.createChooserModel() : null;
         this.selModel = this.parseSelModel(selModel);
-        this.stateModel = this.parseStateModel(stateModel);
+        this.persistenceModel = this.parsePersistenceModel(persistWith, persistOptions);
         this.experimental = this.parseExperimental(experimental);
     }
 
@@ -317,7 +321,7 @@ export class GridModel {
         this.setColumns(columns);
         this.setSortBy(sortBy);
         this.setGroupBy(groupBy);
-        this.stateModel.clear();
+        this.persistenceModel.clear();
     }
 
     /**
@@ -957,18 +961,9 @@ export class GridModel {
         return this.markManaged(new StoreSelectionModel({mode, store: this.store}));
     }
 
-    parseStateModel(val) {
-        if (isPlainObject(val)) {
-            val = {gridModel: this, ...val};
-        } else if (isString(val)) {
-            val = {gridModel: this, provider: {type: 'localStorage', key: val}};
-        } else if (isNil(val)) {
-            val = {gridModel: this, provider: {type: 'transient'}};
-        } else {
-            throw XH.exception(`Could not generate GridStateModel from specification: ${val}`);
-        }
-
-        return new GridStateModel(val);
+    parsePersistenceModel(persistWith, persistOptions) {
+        warnIf(persistOptions && !persistWith, "Must specify 'persistWith' if you wish to persist this grid.");
+        return persistWith ? new PersistenceModel(this, persistWith, persistOptions) : null;
     }
 
     parseExperimental(experimental) {
@@ -1046,3 +1041,10 @@ export class GridModel {
  */
 
 
+/**
+ * @typedef {Object} GridPersistenceOptions
+ * @property {string} [path] - path or key in src where state should be stored (default 'gridModel')
+ * @property {boolean} [persistColumns] - true to include column information (default true)
+ * @property {boolean} [persistGrouping] - true to include grouping information (default true)
+ * @property {boolean} [persistSort] - true to include sorting information (default true)
+ */