+++ desktop/cmp/grid/impl/ColChooser.js
@@ -9,7 +9,7 @@ import {hoistCmp, uses} from '@xh/hoist/core';
 import {button} from '@xh/hoist/desktop/cmp/button';
 import {leftRightChooser, leftRightChooserFilter} from '@xh/hoist/desktop/cmp/leftrightchooser';
 import {panel} from '@xh/hoist/desktop/cmp/panel';
-import {toolbar, toolbarSep} from '@xh/hoist/desktop/cmp/toolbar';
+import {toolbar} from '@xh/hoist/desktop/cmp/toolbar';
 import {Icon} from '@xh/hoist/icon';
 import {withDefault} from '@xh/hoist/utils/js';
 import {ColChooserModel} from './ColChooserModel';
@@ -32,7 +32,7 @@ export const colChooser = hoistCmp.factory({
     className: 'xh-col-chooser',
 
     render({model, className, width, height}) {
-        const {gridModel, isPopoverOpen} = model;
+        const {isPopoverOpen} = model;
 
         return panel({
             className,
@@ -48,9 +48,6 @@ export const colChooser = hoistCmp.factory({
                         text: 'Reset',
                         icon: Icon.undo({className: 'xh-red'}),
                         onClick: () => model.restoreDefaults()
-                    }),
-                    toolbarSep({
-                        omit: !gridModel.stateModel
                     }),
                     button({
                         text: isPopoverOpen ? 'Close' : 'Cancel',